requirements.txt
PySide6>=6.5.0
PyPDF2>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.21.0
ebooklib>=0.20
beautifulsoup4>=4.12.0